+++ engine/src/starlight/renderer/views/SkyboxRenderView.hh
@@ -25,7 +25,7 @@ public:
 
     void render(
       RendererBackend& renderer, RenderPacket& packet, const RenderProperties& props,
-      float deltaTime, CommandBuffer& commandBuffer, u8 imageIndex
+      float deltaTime, CommandBuffer& commandBuffer, u32 imageIndex
     ) override;
 
 private:
